refactor: rename dirs to conventional names
