+++ backend/api/v1/aro/auth/manual/register.py
@@ -6,7 +6,7 @@ from api.v1.aro.auth.manual.password import (
     verify_password,
 )
 from api.v1.aro.auth.services.tokens import create_auth_token
-from api.v1.aro.models.auth_requests import LoginRequest, RegisterRequest
+from api.v1.aro.schemas.auth_requests import LoginRequest, RegisterRequest
 from data.data_wrappers.wrappers import (
     AROUserAuthTokenWrapper,
     AROUserLoginWrapper,
