+++ backend/api/v1/aro/auth/services/tokens.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
 from uuid import UUID, uuid4
 
-from api.v1.aro.models.auth_requests import GoogleRequest
+from api.v1.aro.schemas.auth_requests import GoogleRequest
 from data.data_wrappers.wrappers import (
     AROUserAuthTokenWrapper,
     AROUsersWrapper,
