+++ backend/api/v1/mcc/routes/sessions.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Annotated
 
-from api.v1.mcc.models.responses import CommsSessionsResponse
+from api.v1.mcc.schemas.responses import CommsSessionsResponse
 from data.data_wrappers.wrappers import CommsSessionWrapper
 from fastapi import APIRouter, Query
 from mcc_keycloak.client import keycloak
