+++ backend/api/v1/mcc/schemas/responses.py
